New default output mixin, many other changes
The new default mixin is the DXsM mixin. This is going
to be the new default in the C++ library and is designed to
be harder to reverse engineer the internal state.

Changed how serialization works, can serialize any type but
only deserialize into a SetSeq type (as that is the only one which
makes sense). Serialization is done with an intermediary struct
which records useful information and verifys it when being
restored to a PCG. For example this prevents an MCG state being
restored to an LCG based PCG.